change: pagination li styles
