+++ CHANGELOG.mdx
@@ -77,6 +77,7 @@ Prefix the change with one of these keywords:
 ### Added
 
 - `Embed`: `isRound` prop that adds rounded-xl class
+- `Pagination`: added margin-y-0 on li elements
 
 ### Changed
